chore: ajustar validações do campo budget para tipo integer em StoreCostCenterRequest e UpdateCostCenterRequest

+++ app/Http/Requests/UpdateCostCenterRequest.php
@@ -14,7 +14,7 @@ class UpdateCostCenterRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'budget' => ['nullable', 'numeric'],
+            'budget' => ['nullable', 'int'],
             'start_date' => ['nullable', 'date'],
             'expected_end_date' => ['nullable', 'date', 'after_or_equal:start_date'],
         ];
